bootstrap_fedora: remove yocto dependencies
I'm only building in containers, so remove these. On a fresh install, I
will probably miss some packages, but those can be added to the list of
packages to install. Also, down latest release of Ubuntu Mono Nerd font,
and add a comment to disable the audit framework from Linux.