shell: env: check if running X11 in .zshrc and .bashrc
Since shell/env is sourced in .zshenv, this check was always false,
because .zshenv is read before X11 is started. From ArchWiki:

Used for setting user's environment variables; it should not contain
commands that produce output or assume the shell is attached to a TTY.
When this file exists it will _always_ be read.