neomutt: add signature
This will insert whatever is in $signature after "-- \n" (note the
trailing space). These dashes are enabled by default via sig_dahes, and
it is strongly recommended to not disable them. The reason for this is
because many software packages use "-- \n" to detect your signature.
For example, NeoMutt has the ability to highlight the signature in a
different color in the built-in pager.