git: sort tags by -rc
With the default sorting of tags, the tags will be out of order in, for
example, the Linux repository:

    $ git tag -l
    v6.11
    v6.11-rc1
    v6.11-rc2
    v6.11-rc3
    v6.11-rc4
    v6.11-rc5
    v6.11-rc6
    v6.11-rc7
    v6.11.1
    v6.11.10
    v6.11.2

After this commit, the tags will be correctly ordered like so:

    $ git tag -l
    v6.11-rc1
    v6.11-rc2
    v6.11-rc3
    v6.11-rc4
    v6.11-rc5
    v6.11-rc6
    v6.11-rc7
    v6.11
    v6.11.1
    v6.11.2

Also, the tags will be shown in the correct order from oldest to newest.
For example, the latest 3 tags was

    v6.9.7
    v6.9.8
    v6.9.9

before this commit, while they are

    v6.12-rc7
    v6.12
    v6.12.1

after this commit.