rofi: set transparent border color for listview
After updating to rofi-wayland version 1.7.7+wayland1, an annoying
gray box/line appeared under the input box. It turns out that the
listview widget inherits the opacity setting from somewhere to the
border-color variable. To make the gray box/line disappear, either set
the opacity to 0% or use the transparent color. This color can
probably also be used elsewhere in the configuration.