alacritty: clean up the colors section
This commit mostly removes a bunch of unnecessary or unused settings:

- Remove indexed_colors. The documentation reads: "When these are not
  set, they're filled with sensible defaults." and I don't see any
  difference after removing the indexed colors, so it seems sensible to
  remove them.

- Remove dim colors. Like the indexed colors, setting the dim colors
  doesn't seem to make any difference, so remove them.

- Remove cursor, vi_mode_cursor, and selection. These blocks were all
  set to their default values.

- Remove the dim_foreground and bright_foreground attributes from the
  primary block.

- Remove dim_foreground, which "[...] is automatically calculated based
  on the foreground color.", from the primary block. Also, remove
  bright_foreground because it is only used when
  draw_bold_text_with_bright_colors is set, which I don't.

Furthermore, the primary, normal, and bright color blocks have been
sorted to be consistent with the documentation, that is, black, red,
green, ...

As a side note, the colors are now almost identical to the official
Gruvbox theme [1] except that the normal and bright blocks are swapped
except for black. If black is not swapped, the symlink arrow (->) and
permission line (-) from eza becomes invisible. Using the bright colors
as normal makes the colors consistent with Neovim and other terminal
applications.

[1]: https://github.com/alacritty/alacritty-theme/blob/master/themes/gruvbox_dark.toml