nvim: lsp: migrate to 0.11
After the introduction of vim.lsp.config() and vim.lsp.enable(), the
configuration can be simplified. At this point, the nvim-lspconfig
plugin just acts as a collection of configuration files for language
servers. Note that bitbake_ls does not yet work with the new setup [1],
so it has been commented for now.

When using the 0.11+ setup, it is no longer necessary to add
capabilities from blink [2, 3]. Consequently, the dependency on
blink.cmp can be dropped. This works because nvim-lspconfig is always
loaded before blink.cmp and thus the language servers are available when
blink loads.

Also, drop unused (i.e, commented) servers.

[1]: https://github.com/neovim/nvim-lspconfig/issues/3705
[2]: https://github.com/Saghen/blink.cmp/blob/main/plugin/blink-cmp.lua
[3]: https://cmp.saghen.dev/installation