nvim: conform: add ansible-lint
Running ansible-lint is _really_ slow, so a timeout is necessary to make
conform happy. This timeout could also have been set globally for all
formatters via

    require("conform").format({ timeout_ms = 3000 })

in the keys mapping. However, I want to limit it to only ansible-lint,
so this is just a note to myself. Another option would have been to make
to formatting asynchronous via

    require("conform").format({ async = true }),

but I want a to set some upper limit in case ansible-lint becomes even
slower.

Also, drop setting the mode for formatting as this is always done in
normal mode.