nvim: lualine: drop useless information
The lualine_x section contains the encoding, fileformat, and filetype
components by default. I cannot recall when the last time was I worked
on a file that was not UTF-8 encoded and in Unix format. Icons from the
filetype component are nice, and I will re-add those in the following
commit.